perform image transforms using Transformable interface
